Dispositions Data

The disposition data presented in this data reflects the culmination of the fact-finding process that leads to the resolution of a case. Each row represents a charge that has been disposed of.

dispositions

Format

The data frame dispositions has 20,234 rows and 31 variables:

case_id

Internal unique identifier for each case

case_participant_id

Internal unique identifier for each person associated with a case

offense_category

Broad offense categories before specific charges are filed on a case

primary_charge

A flag for the top charge, usually the way the case is referred to

charge_id

Internal unique identifier for each charge filed

charge_version_id

Internal unique identifier for each version of a charge associated with charges filed

disposition_charged_offense_title

The specific title of the charged offense at the disposition

disposition_charged_chapter

The legal chapter for the charge at the disposition

disposition_charged_act

The legal act for the charge at the disposition

disposition_charged_section

The legal section for the charge at the disposition

disposition_charged_class

The legal class for the charge at the disposition

disposition_aoic

Administrative Office of the Illinois Courts ID for law of the charge at the disposition

disposition_date

The date the charge was disposed of

charge_disposition

The result of the charge

charge_disposition_reason

Additional information about the result of the charge

judge

Judge who oversaw the case

court_name

The Circuit Court District the charge was disposed of in

court_facility

The courthouse the charge was disposed of in

age_at_incident

Recorded age at the time of the incident

gender

Recorded gender of the defendant

race

Recorded race of the defendant

incident_begin_date

Date of when the incident began

incident_end_date

Date of when the incident ended (this will be blank for incidents that did not go more than one day)

arrest_date

Date and time of arrest

law_enforcement_agency

Law enforcement agency associated with the arrest

unit

The law enforcement unit associated with the arrest

incident_city

The city where the incident took place

received_date

Date when felony review received the case

arraignment_date

Date of the arraignment

update_offense_category

This field is the offense category for the case updated based upon the top charge for the primary offender. It can differ from the first offense category assigned to the case in part because cases evolve.

charge_count

The charge count of the charged offense

Source

https://datacatalog.cookcountyil.gov/Courts/Initiation/7mck-ehwz

See also

Examples

str(dispositions)

#> 'data.frame':	20234 obs. of  31 variables:
#>  $ case_id                          : chr  "129568558770" "129568602572" "129568602572" "129568602572" ...
#>  $ case_participant_id              : chr  "1079960508495" "1079960625753" "1079960625753" "1079960625753" ...
#>  $ offense_category                 : chr  "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" ...
#>  $ primary_charge                   : chr  "true" "false" "false" "false" ...
#>  $ charge_id                        : chr  "2534066046185" "2536166541538" "2536167428821" "2536207948093" ...
#>  $ charge_version_id                : chr  "577689205836" "578179378084" "578179579469" "578189111715" ...
#>  $ disposition_charged_offense_title: chr  "RECKLESS DISCHARGE OF A FIREARM" "UNLAWFUL USE OR POSSESSION OF A WEAPON BY A FELON" "UNLAWFUL USE OR POSSESSION OF A WEAPON BY A FELON" "AGGRAVATED UNLAWFUL USE OF WEAPON" ...
#>  $ disposition_charged_chapter      : chr  "720" "720" "720" "720" ...
#>  $ disposition_charged_act          : chr  "5" "5" "5" "5" ...
#>  $ disposition_charged_section      : chr  "24-1.5(a)" "24-1.1(a)" "24-1.1(a)" "24-1.6(a)(1)" ...
#>  $ disposition_charged_class        : chr  "4" "2" "2" "2" ...
#>  $ disposition_charged_aoic         : chr  "0010322" "0012309" "0012310" "0012475" ...
#>  $ disposition_date                 : chr  "4/8/2019 12:00:00 AM" "8/5/2019 12:00:00 AM" "8/5/2019 12:00:00 AM" "8/5/2019 12:00:00 AM" ...
#>  $ charge_disposition               : chr  "Plea Of Guilty" "Nolle Prosecution" "Nolle Prosecution" "Nolle Prosecution" ...
#>  $ charge_disposition_reason        : chr  "" "PG to Other Count/s" "PG to Other Count/s" "PG to Other Count/s" ...
#>  $ judge                            : chr  "Thaddeus L Wilson" "Thaddeus L Wilson" "Thaddeus L Wilson" "Thaddeus L Wilson" ...
#>  $ court_name                       : chr  "District 1 - Chicago" "District 1 - Chicago" "District 1 - Chicago" "District 1 - Chicago" ...
#>  $ court_facility                   : chr  "26TH Street" "26TH Street" "26TH Street" "26TH Street" ...
#>  $ age_at_incident                  : int  42 20 20 20 20 20 20 20 20 20 ...
#>  $ gender                           : chr  "Male" "Male" "Male" "Male" ...
#>  $ race                             : chr  "Black" "White [Hispanic or Latino]" "White [Hispanic or Latino]" "White [Hispanic or Latino]" ...
#>  $ incident_begin_date              : POSIXct, format: "2018-12-31" "2019-01-01" ...
#>  $ incident_end_date                : POSIXct, format: NA NA ...
#>  $ arrest_date                      : POSIXct, format: "2018-12-31 22:55:00" "2018-12-31 23:28:00" ...
#>  $ law_enforcement_agency           : chr  "CHICAGO PD" "CHICAGO PD" "CHICAGO PD" "CHICAGO PD" ...
#>  $ unit                             : chr  "District 10 - Ogden" "District 10 - Ogden" "District 10 - Ogden" "District 10 - Ogden" ...
#>  $ incident_city                    : chr  "Chicago" "Chicago" "Chicago" "Chicago" ...
#>  $ received_date                    : POSIXct, format: "2019-01-01" "2019-01-01" ...
#>  $ arraignment_date                 : POSIXct, format: "2019-02-05" "2019-02-07" ...
#>  $ updated_offense_category         : chr  "Reckless Discharge of Firearm" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" ...
#>  $ charge_count                     : int  1 2 3 4 5 6 7 8 9 10 ...

head(dispositions)

#>        case_id case_participant_id             offense_category primary_charge
#> 1 129568558770       1079960508495 UUW - Unlawful Use of Weapon           true
#> 2 129568602572       1079960625753 UUW - Unlawful Use of Weapon          false
#> 3 129568602572       1079960625753 UUW - Unlawful Use of Weapon          false
#> 4 129568602572       1079960625753 UUW - Unlawful Use of Weapon          false
#> 5 129568602572       1079960625753 UUW - Unlawful Use of Weapon          false
#> 6 129568602572       1079960625753 UUW - Unlawful Use of Weapon          false
#>       charge_id charge_version_id
#> 1 2534066046185      577689205836
#> 2 2536166541538      578179378084
#> 3 2536167428821      578179579469
#> 4 2536207948093      578189111715
#> 5 2536208539615      578189245972
#> 6 2536209131137      578189380229
#>                   disposition_charged_offense_title disposition_charged_chapter
#> 1                   RECKLESS DISCHARGE OF A FIREARM                         720
#> 2 UNLAWFUL USE OR POSSESSION OF A WEAPON BY A FELON                         720
#> 3 UNLAWFUL USE OR POSSESSION OF A WEAPON BY A FELON                         720
#> 4                 AGGRAVATED UNLAWFUL USE OF WEAPON                         720
#> 5                 AGGRAVATED UNLAWFUL USE OF WEAPON                         720
#> 6                 AGGRAVATED UNLAWFUL USE OF WEAPON                         720
#>   disposition_charged_act disposition_charged_section disposition_charged_class
#> 1                       5                   24-1.5(a)                         4
#> 2                       5                   24-1.1(a)                         2
#> 3                       5                   24-1.1(a)                         2
#> 4                       5                24-1.6(a)(1)                         2
#> 5                       5                24-1.6(a)(1)                         2
#> 6                       5                24-1.6(a)(1)                         2
#>   disposition_charged_aoic     disposition_date charge_disposition
#> 1                  0010322 4/8/2019 12:00:00 AM     Plea Of Guilty
#> 2                  0012309 8/5/2019 12:00:00 AM  Nolle Prosecution
#> 3                  0012310 8/5/2019 12:00:00 AM  Nolle Prosecution
#> 4                  0012475 8/5/2019 12:00:00 AM  Nolle Prosecution
#> 5                  0012475 8/5/2019 12:00:00 AM  Nolle Prosecution
#> 6                  0012475 8/5/2019 12:00:00 AM  Nolle Prosecution
#>   charge_disposition_reason             judge           court_name
#> 1                           Thaddeus L Wilson District 1 - Chicago
#> 2       PG to Other Count/s Thaddeus L Wilson District 1 - Chicago
#> 3       PG to Other Count/s Thaddeus L Wilson District 1 - Chicago
#> 4       PG to Other Count/s Thaddeus L Wilson District 1 - Chicago
#> 5       PG to Other Count/s Thaddeus L Wilson District 1 - Chicago
#> 6       PG to Other Count/s Thaddeus L Wilson District 1 - Chicago
#>   court_facility age_at_incident gender                       race
#> 1    26TH Street              42   Male                      Black
#> 2    26TH Street              20   Male White [Hispanic or Latino]
#> 3    26TH Street              20   Male White [Hispanic or Latino]
#> 4    26TH Street              20   Male White [Hispanic or Latino]
#> 5    26TH Street              20   Male White [Hispanic or Latino]
#> 6    26TH Street              20   Male White [Hispanic or Latino]
#>   incident_begin_date incident_end_date         arrest_date
#> 1          2018-12-31              <NA> 2018-12-31 22:55:00
#> 2          2019-01-01              <NA> 2018-12-31 23:28:00
#> 3          2019-01-01              <NA> 2018-12-31 23:28:00
#> 4          2019-01-01              <NA> 2018-12-31 23:28:00
#> 5          2019-01-01              <NA> 2018-12-31 23:28:00
#> 6          2019-01-01              <NA> 2018-12-31 23:28:00
#>   law_enforcement_agency                unit incident_city received_date
#> 1             CHICAGO PD District 10 - Ogden       Chicago    2019-01-01
#> 2             CHICAGO PD District 10 - Ogden       Chicago    2019-01-01
#> 3             CHICAGO PD District 10 - Ogden       Chicago    2019-01-01
#> 4             CHICAGO PD District 10 - Ogden       Chicago    2019-01-01
#> 5             CHICAGO PD District 10 - Ogden       Chicago    2019-01-01
#> 6             CHICAGO PD District 10 - Ogden       Chicago    2019-01-01
#>   arraignment_date      updated_offense_category charge_count
#> 1       2019-02-05 Reckless Discharge of Firearm            1
#> 2       2019-02-07  UUW - Unlawful Use of Weapon            2
#> 3       2019-02-07  UUW - Unlawful Use of Weapon            3
#> 4       2019-02-07  UUW - Unlawful Use of Weapon            4
#> 5       2019-02-07  UUW - Unlawful Use of Weapon            5
#> 6       2019-02-07  UUW - Unlawful Use of Weapon            6

summary(dispositions)

#>    case_id          case_participant_id offense_category   primary_charge    
#>  Length:20234       Length:20234        Length:20234       Length:20234      
#>  Class :character   Class :character    Class :character   Class :character  
#>  Mode  :character   Mode  :character    Mode  :character   Mode  :character  
#>                                                                              
#>                                                                              
#>                                                                              
#>                                                                              
#>   charge_id         charge_version_id  disposition_charged_offense_title
#>  Length:20234       Length:20234       Length:20234                     
#>  Class :character   Class :character   Class :character                 
#>  Mode  :character   Mode  :character   Mode  :character                 
#>                                                                         
#>                                                                         
#>                                                                         
#>                                                                         
#>  disposition_charged_chapter disposition_charged_act
#>  Length:20234                Length:20234           
#>  Class :character            Class :character       
#>  Mode  :character            Mode  :character       
#>                                                     
#>                                                     
#>                                                     
#>                                                     
#>  disposition_charged_section disposition_charged_class disposition_charged_aoic
#>  Length:20234                Length:20234              Length:20234            
#>  Class :character            Class :character          Class :character        
#>  Mode  :character            Mode  :character          Mode  :character        
#>                                                                                
#>                                                                                
#>                                                                                
#>                                                                                
#>  disposition_date   charge_disposition charge_disposition_reason
#>  Length:20234       Length:20234       Length:20234             
#>  Class :character   Class :character   Class :character         
#>  Mode  :character   Mode  :character   Mode  :character         
#>                                                                 
#>                                                                 
#>                                                                 
#>                                                                 
#>     judge            court_name        court_facility     age_at_incident 
#>  Length:20234       Length:20234       Length:20234       Min.   : 18.00  
#>  Class :character   Class :character   Class :character   1st Qu.: 23.00  
#>  Mode  :character   Mode  :character   Mode  :character   Median : 30.00  
#>                                                           Mean   : 33.61  
#>                                                           3rd Qu.: 42.00  
#>                                                           Max.   :119.00  
#>                                                           NA's   :295     
#>     gender              race           incident_begin_date          
#>  Length:20234       Length:20234       Min.   :1998-10-17 00:00:00  
#>  Class :character   Class :character   1st Qu.:2019-02-03 00:00:00  
#>  Mode  :character   Mode  :character   Median :2019-03-30 00:00:00  
#>                                        Mean   :2019-03-28 10:44:42  
#>                                        3rd Qu.:2019-06-09 00:00:00  
#>                                        Max.   :2019-11-15 00:00:00  
#>                                        NA's   :275                  
#>  incident_end_date              arrest_date                 
#>  Min.   :2001-06-26 00:00:00   Min.   :1965-11-06 15:59:00  
#>  1st Qu.:2019-01-11 00:00:00   1st Qu.:2019-02-12 12:21:00  
#>  Median :2019-03-08 00:00:00   Median :2019-04-05 19:14:00  
#>  Mean   :2019-01-20 18:29:27   Mean   :2019-04-13 13:45:28  
#>  3rd Qu.:2019-06-23 12:00:00   3rd Qu.:2019-06-13 08:14:00  
#>  Max.   :2019-11-11 00:00:00   Max.   :2019-11-15 21:44:00  
#>  NA's   :19354                 NA's   :551                  
#>  law_enforcement_agency     unit           incident_city     
#>  Length:20234           Length:20234       Length:20234      
#>  Class :character       Class :character   Class :character  
#>  Mode  :character       Mode  :character   Mode  :character  
#>                                                              
#>                                                              
#>                                                              
#>                                                              
#>  received_date                 arraignment_date             
#>  Min.   :2019-01-01 00:00:00   Min.   :2018-02-01 00:00:00  
#>  1st Qu.:2019-02-14 00:00:00   1st Qu.:2019-03-21 00:00:00  
#>  Median :2019-04-07 00:00:00   Median :2019-05-08 00:00:00  
#>  Mean   :2019-04-21 02:28:44   Mean   :2019-05-15 22:13:24  
#>  3rd Qu.:2019-06-18 00:00:00   3rd Qu.:2019-07-03 00:00:00  
#>  Max.   :2019-11-22 00:00:00   Max.   :2019-12-06 00:00:00  
#>                                NA's   :6658                 
#>  updated_offense_category  charge_count   
#>  Length:20234             Min.   : 1.000  
#>  Class :character         1st Qu.: 1.000  
#>  Mode  :character         Median : 1.000  
#>                           Mean   : 2.837  
#>                           3rd Qu.: 3.000  
#>                           Max.   :53.000  
#>