Initiation Data

The Initiation results data presented here reflects all of the arrests that came through the door of the State's Attorneys Office (SAO). An initiation is how an arrest turns into a “case” in the courts. Most cases are initiated through a process known as felony review, in which SAO attorneys make a decision whether or not to prosecute. Cases may also be indicted by a grand jury or, in narcotics cases, filed directly by law enforcement (labeled "BOND SET (Narcotics)" in this data). Included in this data set are the defendant counts by initiation and year. This data includes felony cases handled by the Criminal, Narcotics, and Special Prosecution Bureaus. It does not include information about cases processed through the Juvenile Justice and Civil Actions Bureaus.

initiation

Format

The data frame intake has 73,029 rows and 27 variables:

case_id

Internal unique identifier for each case

case_participant_id

Internal unique identifier for each person associated with a case

offense_category

Broad offense categories before specific charges are filed on a case

primary_charge

A flag for the top charge, usually the way the case is referred to

charge_id

Internal unique identifier for each charge filed

charge_version_id

Internal unique identifier for each version of a charge associated with charges filed

charge_offense_title

The offense title of the charged offense

chapter

The legal chapter for the charge

act

The legal act for the charge

section

The legal section for the charge

class

The legal class for the charge

aoic

Administrative Office of the Illinois Courts ID for law of the charge

event

The way the charge was brought about

event_date

The date the charges were brought about

age_at_incident

Recorded age at the time of the incident

gender

Recorded gender of the defendant

race

Recorded race of the defendant

incident_begin_date

Date of when the incident began

incident_end_date

Date of when the incident ended (this will be blank for incidents that did not go more than one day)

arrest_date

Date and time of arrest

law_enforcement_agency

Law enforcement agency associated with the arrest

unit

The law enforcement unit associated with the arrest

incident_city

The city where the incident took place

received_date

Date when felony review received the case

arraignment_date

Date of the arraignment

update_offense_category

This field is the offense category for the case updated based upon the top charge for the primary offender. It can differ from the first offense category assigned to the case in part because cases evolve.

charge_count

The charge count of the charged offense

Source

https://datacatalog.cookcountyil.gov/Courts/Initiation/7mck-ehwz

See also

Examples

str(initiation)

#> 'data.frame':	73029 obs. of  27 variables:
#>  $ case_id                 : chr  "129405877581" "129568427364" "129568471166" "129568471166" ...
#>  $ case_participant_id     : chr  "1077995851303" "1079958749626" "1079959687690" "1079959687690" ...
#>  $ offense_category        : chr  "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" ...
#>  $ primary_charge          : chr  "true" "true" "false" "false" ...
#>  $ charge_id               : chr  "2529188058314" "2534063384335" "2534433381475" "2534434268758" ...
#>  $ charge_version_id       : chr  "576544933565" "577688601680" "577775465948" "577775667334" ...
#>  $ charge_offense_title    : chr  "[POSSESSION OF CONTROLLED SUBSTANCE WITH INTENT TO DELIVER/DELIVERY OF A CONTROLLED SUBSTANCE]" "UNLAWFUL USE OR POSSESSION OF A WEAPON BY A FELON" "AGG UUW/LOADED/NO FCCA/FOID" "AGGRAVATED UNLAWFUL USE OF WEAPON" ...
#>  $ chapter                 : chr  "720" "720" "720" "720" ...
#>  $ act                     : chr  "570" "5" "5" "5" ...
#>  $ section                 : chr  "401(c)(2)" "24-1.1(a)" "24-1.6(a)(1)" "24-1.6(a)(1)" ...
#>  $ class                   : chr  "1" "2" "4" "4" ...
#>  $ aoic                    : chr  "5096200" "0012309" "0017785" "0017862" ...
#>  $ event                   : chr  "Indictment" "Preliminary Hearing" "Preliminary Hearing" "Preliminary Hearing" ...
#>  $ event_date              : POSIXct, format: "2019-06-06" "2019-01-15" ...
#>  $ age_at_incident         : int  31 24 20 20 20 20 20 20 20 20 ...
#>  $ gender                  : chr  "Male" "Male" "Male" "Male" ...
#>  $ race                    : chr  "Black" "Black" "Black" "Black" ...
#>  $ incident_begin_date     : POSIXct, format: "2018-11-12" "2019-01-01" ...
#>  $ incident_end_date       : POSIXct, format: NA NA ...
#>  $ arrest_date             : POSIXct, format: "2018-11-12 23:23:00" NA ...
#>  $ law_enforcement_agency  : chr  "CHICAGO PD" "CHICAGO PD" "CHICAGO PD" "CHICAGO PD" ...
#>  $ unit                    : chr  "District 10 - Ogden" "District 3 - Grand Crossing" "District 22 - Morgan Park" "District 22 - Morgan Park" ...
#>  $ incident_city           : chr  "Chicago" "Chicago" "Chicago" "Chicago" ...
#>  $ received_date           : POSIXct, format: "2019-01-05" "2019-01-01" ...
#>  $ arraignment_date        : POSIXct, format: "2019-06-20" NA ...
#>  $ updated_offense_category: chr  "Narcotics" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" ...
#>  $ charge_count            : int  1 1 2 3 4 5 6 7 8 1 ...

head(initiation)

#>        case_id case_participant_id             offense_category primary_charge
#> 1 129405877581       1077995851303 UUW - Unlawful Use of Weapon           true
#> 2 129568427364       1079958749626 UUW - Unlawful Use of Weapon           true
#> 3 129568471166       1079959687690 UUW - Unlawful Use of Weapon          false
#> 4 129568471166       1079959687690 UUW - Unlawful Use of Weapon          false
#> 5 129568471166       1079959687690 UUW - Unlawful Use of Weapon          false
#> 6 129568471166       1079959687690 UUW - Unlawful Use of Weapon          false
#>       charge_id charge_version_id
#> 1 2529188058314      576544933565
#> 2 2534063384335      577688601680
#> 3 2534433381475      577775465948
#> 4 2534434268758      577775667334
#> 5 2534434860280      577775801591
#> 6 2534435156042      577775868719
#>                                                                             charge_offense_title
#> 1 [POSSESSION OF CONTROLLED SUBSTANCE WITH INTENT TO DELIVER/DELIVERY OF A CONTROLLED SUBSTANCE]
#> 2                                              UNLAWFUL USE OR POSSESSION OF A WEAPON BY A FELON
#> 3                                                                    AGG UUW/LOADED/NO FCCA/FOID
#> 4                                                              AGGRAVATED UNLAWFUL USE OF WEAPON
#> 5                                                              AGGRAVATED UNLAWFUL USE OF WEAPON
#> 6                                                              AGGRAVATED UNLAWFUL USE OF WEAPON
#>   chapter act      section class    aoic               event event_date
#> 1     720 570    401(c)(2)     1 5096200          Indictment 2019-06-06
#> 2     720   5    24-1.1(a)     2 0012309 Preliminary Hearing 2019-01-15
#> 3     720   5 24-1.6(a)(1)     4 0017785 Preliminary Hearing 2019-02-01
#> 4     720   5 24-1.6(a)(1)     4 0017862 Preliminary Hearing 2019-02-01
#> 5     720   5 24-1.6(a)(1)     4 0017864 Preliminary Hearing 2019-02-01
#> 6     720   5 24-1.6(a)(1)     4 0017869 Preliminary Hearing 2019-02-01
#>   age_at_incident gender  race incident_begin_date incident_end_date
#> 1              31   Male Black          2018-11-12              <NA>
#> 2              24   Male Black          2019-01-01              <NA>
#> 3              20   Male Black          2018-12-31              <NA>
#> 4              20   Male Black          2018-12-31              <NA>
#> 5              20   Male Black          2018-12-31              <NA>
#> 6              20   Male Black          2018-12-31              <NA>
#>           arrest_date law_enforcement_agency                        unit
#> 1 2018-11-12 23:23:00             CHICAGO PD         District 10 - Ogden
#> 2                <NA>             CHICAGO PD District 3 - Grand Crossing
#> 3 2018-12-31 22:45:00             CHICAGO PD   District 22 - Morgan Park
#> 4 2018-12-31 22:45:00             CHICAGO PD   District 22 - Morgan Park
#> 5 2018-12-31 22:45:00             CHICAGO PD   District 22 - Morgan Park
#> 6 2018-12-31 22:45:00             CHICAGO PD   District 22 - Morgan Park
#>   incident_city received_date arraignment_date     updated_offense_category
#> 1       Chicago    2019-01-05       2019-06-20                    Narcotics
#> 2       Chicago    2019-01-01             <NA> UUW - Unlawful Use of Weapon
#> 3       Chicago    2019-01-01       2019-02-06 UUW - Unlawful Use of Weapon
#> 4       Chicago    2019-01-01       2019-02-06 UUW - Unlawful Use of Weapon
#> 5       Chicago    2019-01-01       2019-02-06 UUW - Unlawful Use of Weapon
#> 6       Chicago    2019-01-01       2019-02-06 UUW - Unlawful Use of Weapon
#>   charge_count
#> 1            1
#> 2            1
#> 3            2
#> 4            3
#> 5            4
#> 6            5

summary(initiation)

#>    case_id          case_participant_id offense_category   primary_charge    
#>  Length:73029       Length:73029        Length:73029       Length:73029      
#>  Class :character   Class :character    Class :character   Class :character  
#>  Mode  :character   Mode  :character    Mode  :character   Mode  :character  
#>                                                                              
#>                                                                              
#>                                                                              
#>                                                                              
#>   charge_id         charge_version_id  charge_offense_title   chapter         
#>  Length:73029       Length:73029       Length:73029         Length:73029      
#>  Class :character   Class :character   Class :character     Class :character  
#>  Mode  :character   Mode  :character   Mode  :character     Mode  :character  
#>                                                                               
#>                                                                               
#>                                                                               
#>                                                                               
#>      act              section             class               aoic          
#>  Length:73029       Length:73029       Length:73029       Length:73029      
#>  Class :character   Class :character   Class :character   Class :character  
#>  Mode  :character   Mode  :character   Mode  :character   Mode  :character  
#>                                                                             
#>                                                                             
#>                                                                             
#>                                                                             
#>     event             event_date                  age_at_incident 
#>  Length:73029       Min.   :2016-06-26 00:00:00   Min.   : 17.00  
#>  Class :character   1st Qu.:2019-04-10 00:00:00   1st Qu.: 24.00  
#>  Mode  :character   Median :2019-06-19 00:00:00   Median : 30.00  
#>                     Mean   :2019-06-25 09:00:44   Mean   : 32.86  
#>                     3rd Qu.:2019-09-04 00:00:00   3rd Qu.: 40.00  
#>                     Max.   :2109-05-08 00:00:00   Max.   :137.00  
#>                     NA's   :6472                  NA's   :1685    
#>     gender              race           incident_begin_date          
#>  Length:73029       Length:73029       Min.   :1979-12-31 00:00:00  
#>  Class :character   Class :character   1st Qu.:2019-02-23 00:00:00  
#>  Mode  :character   Mode  :character   Median :2019-05-10 00:00:00  
#>                                        Mean   :2019-03-26 05:04:05  
#>                                        3rd Qu.:2019-08-01 00:00:00  
#>                                        Max.   :2019-12-01 00:00:00  
#>                                        NA's   :1520                 
#>  incident_end_date              arrest_date                 
#>  Min.   :1989-05-24 00:00:00   Min.   :1965-11-06 15:59:00  
#>  1st Qu.:2018-12-06 00:00:00   1st Qu.:2019-03-16 09:46:00  
#>  Median :2019-03-14 00:00:00   Median :2019-05-25 22:07:00  
#>  Mean   :2018-07-10 11:22:11   Mean   :2019-05-27 01:00:05  
#>  3rd Qu.:2019-06-26 00:00:00   3rd Qu.:2019-08-10 04:44:00  
#>  Max.   :2019-11-28 00:00:00   Max.   :2019-12-01 23:25:00  
#>  NA's   :68134                 NA's   :3288                 
#>  law_enforcement_agency     unit           incident_city     
#>  Length:73029           Length:73029       Length:73029      
#>  Class :character       Class :character   Class :character  
#>  Mode  :character       Mode  :character   Mode  :character  
#>                                                              
#>                                                              
#>                                                              
#>                                                              
#>  received_date                 arraignment_date             
#>  Min.   :2019-01-01 00:00:00   Min.   :2009-09-19 00:00:00  
#>  1st Qu.:2019-03-17 00:00:00   1st Qu.:2019-04-29 00:00:00  
#>  Median :2019-05-26 00:00:00   Median :2019-07-05 00:00:00  
#>  Mean   :2019-06-01 06:09:33   Mean   :2019-07-09 03:03:01  
#>  3rd Qu.:2019-08-14 00:00:00   3rd Qu.:2019-09-17 00:00:00  
#>  Max.   :2019-12-02 00:00:00   Max.   :2109-07-24 00:00:00  
#>                                NA's   :13795                
#>  updated_offense_category  charge_count   
#>  Length:73029             Min.   : 1.000  
#>  Class :character         1st Qu.: 1.000  
#>  Mode  :character         Median : 2.000  
#>                           Mean   : 4.094  
#>                           3rd Qu.: 5.000  
#>                           Max.   :82.000  
#>