Intake Data

The intake data presented in this data reflects the cases brought in for review. Each row represents a potential defendant in a case.

intake

Format

The data frame intake has 36,190 rows and 17 variables:

case_id

Internal unique identifier for each case

case_participant_id

Internal unique identifier for each person associated with a case

received_date

Date when felony review received the case

offense_category

Broad offense categories before specific charges are filed on a case

participant_status

The status of a case brought against a person. This status does not necessarily come from Felony Review

age_at_incident

Recorded age at the time of the incident

gender

Recorded gender of the defendant

race

Recorded race of the defendant

incident_begin_date

Date of when the incident began

incident_end_date

Date of when the incident ended (this will be blank for incidents that did not go more than one day)

law_enforcement_agency

Law enforcement agency associated with the arrest

unit

The law enforcement unit associated with the arrest

incident_city

The city where the incident took place

arrest_date

Date and time of arrest

felony_review_date

The date of the felony review process

felony_review_result

The result of the felony review process for cases that went through felony review

update_offense_category

This field is the offense category for the case updated based upon the top charge for the primary offender. It can differ from the first offense category assigned to the case in part because cases evolve.

Source

https://datacatalog.cookcountyil.gov/Courts/Intake/3k7z-hchi

See also

Examples

str(intake)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	36190 obs. of  17 variables:
#>  $ case_id                : chr  "129405877581" "129568427364" "129568471166" "129568514968" ...
#>  $ case_participant_id    : chr  "1077995851303" "1079958749626" "1079959687690" "1079959922206" ...
#>  $ received_date          : POSIXct, format: "2019-01-05" "2019-01-01" ...
#>  $ offense_category       : Factor w/ 87 levels "Narcotics","Fraud",..: 8 8 8 45 8 8 13 8 8 8 ...
#>  $ participant_status     : Factor w/ 19 levels "Approved","Defendant Dispo'd/See Related Case",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ age_at_incident        : int  31 24 20 35 42 20 30 35 31 NA ...
#>  $ gender                 : Factor w/ 6 levels "Male","Female",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ race                   : Factor w/ 13 levels "White [Hispanic or Latino]",..: 2 2 2 1 2 1 2 2 2 2 ...
#>  $ incident_begin_date    : POSIXct, format: "2018-11-12" "2019-01-01" ...
#>  $ incident_end_date      : POSIXct, format: NA NA ...
#>  $ law_enforcement_agency : Factor w/ 306 levels "CHICAGO PD","IL ATTORNEY GENERAL, CHICAGO",..: 1 1 1 1 1 1 1 1 1 20 ...
#>  $ unit                   : chr  "District 10 - Ogden" "District 3 - Grand Crossing" "District 22 - Morgan Park" "District 4 - South Chicago" ...
#>  $ incident_city          : chr  "Chicago" "Chicago" "Chicago" "Chicago" ...
#>  $ arrest_date            : POSIXct, format: "2018-11-12 23:23:00" NA ...
#>  $ felony_review_date     : POSIXct, format: NA "2019-01-01" ...
#>  $ felony_review_result   : chr  "" "Approved" "Approved" "Continued Investigation" ...
#>  $ update_offense_category: chr  "Narcotics" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "Reckless Discharge of Firearm" ...