Sentencing Data

The sentencing data presented in this report reflects the judgment imposed by the court on people that have been found guilty. Each row represents a charge that has been sentenced.

sentencing

Format

The data frame sentencing has 4,329 rows and 39 variables:

case_id

Internal unique identifier for each case

case_participant_id

Internal unique identifier for each person associated with a case

offense_category

Broad offense categories before specific charges are filed on a case

primary_charge

A flag for the top charge, usually the way the case is referred to

charge_id

Internal unique identifier for each charge filed

charge_version_id

Internal unique identifier for each version of a charge associated with charges filed

disposition_charged_offense_title

The specific title of the charged offense

disposition_charged_chapter

The legal chapter for the charge

disposition_charged_act

The legal act for the charge

disposition_charged_section

The legal section for the charge

disposition_charged_class

The legal class for the charge

disposition_aoic

Administrative Office of the Illinois Courts ID for law of the charge

disposition_date

The date the charge was disposed of

charge_disposition

The result of the charge

charge_disposition_reason

Additional information about the result of the charge

sentence_phase

Sentencing phase explains when this version of the sentence was created

sentence_date

Date of when the charge was sentenced

sentence_judge

Judge who oversaw the sentencing

sentence_type

A broad type of sentence issued

current_sentence

This is a flag which row represents a current sentence

commitment_type

A more specific type of sentence issued

commitment_term

The number associated with the sentence (use this with commitment_unit to understand length of sentence)

commitment_unit

The unit associated with the sentence (use this with commitment_term to understand length of sentence)

court_name

The Circuit Court District the charge was disposed of in

court_facility

The courthouse the charge was disposed of in

length_of_case_in_days

Number of days between a charge being arraigned and a charge being sentenced

age_at_incident

Recorded age at the time of the incident

gender

Recorded gender of the defendant

race

Recorded race of the defendant

incident_begin_date

Date of when the incident began

incident_end_date

Date of when the incident ended (this will be blank for incidents that did not go more than one day)

arrest_date

Date and time of arrest

law_enforcement_agency

Law enforcement agency associated with the arrest

unit

The law enforcement unit associated with the arrest

incident_city

The city where the incident took place

received_date

Date when felony review received the case

arraignment_date

Date of the arraignment

update_offense_category

This field is the offense category for the case updated based upon the top charge for the primary offender. It can differ from the first offense category assigned to the case in part because cases evolve.

charge_count

The charge count of the charged offense

Source

https://datacatalog.cookcountyil.gov/Courts/Sentencing/tg8v-tm6u

See also

Examples

str(sentencing)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	4329 obs. of  39 variables:
#>  $ case_id                          : chr  "129568558770" "129568602572" "129568690177" "129568821583" ...
#>  $ case_participant_id              : chr  "1079960508495" "1079960625753" "1079961798333" "1079962853655" ...
#>  $ offense_category                 : chr  "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "Reckless Discharge of Firearm" ...
#>  $ primary_charge                   : chr  "true" "false" "false" "false" ...
#>  $ charge_id                        : chr  "2534066046185" "2536217412448" "2534064863141" "2534067524991" ...
#>  $ charge_version_id                : chr  "577689205836" "578191662598" "577688937322" "585802221098" ...
#>  $ disposition_charged_offense_title: chr  "RECKLESS DISCHARGE OF A FIREARM" "UNLAWFUL POSSESSION OF A FIREARM BY A STREET GANG MEMBER" "RECKLESS DISCHARGE OF A FIREARM" "UNLAWFUL USE OF A WEAPON" ...
#>  $ disposition_charged_chapter      : chr  "720" "720" "720" "720" ...
#>  $ disposition_charged_act          : chr  "5" "5" "5" "5" ...
#>  $ disposition_charged_section      : chr  "24-1.5(a)" "24-1.8(a)(1)" "24-1.5(a)" "24-1(a)(4)" ...
#>  $ disposition_charged_class        : chr  "4" "2" "4" "A" ...
#>  $ disposition_charged_aoic         : chr  "0010322" "0015061" "0010322" "1210600" ...
#>  $ disposition_date                 : chr  "4/8/2019 12:00:00 AM" "8/5/2019 12:00:00 AM" "11/19/2019 12:00:00 AM" "7/31/2019 12:00:00 AM" ...
#>  $ charge_disposition               : chr  "Plea Of Guilty" "Finding Guilty" "Plea Of Guilty" "Plea Of Guilty" ...
#>  $ charge_disposition_reason        : chr  "" "" "" "" ...
#>  $ sentence_phase                   : chr  "Original Sentencing" "Original Sentencing" "Original Sentencing" "Original Sentencing" ...
#>  $ sentence_date                    : chr  "4/8/2019 12:00:00 AM" "8/5/2019 12:00:00 AM" "11/19/2019 12:00:00 AM" "7/31/2019 12:00:00 AM" ...
#>  $ sentence_judge                   : chr  "Thaddeus L Wilson" "Thaddeus L Wilson" "William G Gamboney" "Catherine Marie Haberkorn" ...
#>  $ sentence_type                    : chr  "Probation" "Prison" "Prison" "Probation" ...
#>  $ current_sentence                 : chr  "true" "true" "true" "true" ...
#>  $ commitment_type                  : chr  "Probation" "Illinois Department of Corrections" "Illinois Department of Corrections" "Probation" ...
#>  $ commitment_term                  : chr  "2" "54" "1" "1" ...
#>  $ commitment_unit                  : chr  "Year(s)" "Months" "Year(s)" "Year(s)" ...
#>  $ court_name                       : chr  "District 1 - Chicago" "District 1 - Chicago" "District 1 - Chicago" "District 2 - Skokie" ...
#>  $ court_facility                   : chr  "26TH Street" "26TH Street" "26TH Street" "Skokie Courthouse" ...
#>  $ length_of_case_in_days           : int  62 179 291 175 92 141 215 33 204 181 ...
#>  $ age_at_incident                  : int  42 20 35 55 21 34 67 31 28 39 ...
#>  $ gender                           : Factor w/ 6 levels "Male","Female",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ race                             : Factor w/ 11 levels "Black","White [Hispanic or Latino]",..: 1 2 1 1 1 2 1 1 1 2 ...
#>  $ incident_begin_date              : POSIXct, format: "2018-12-31" "2019-01-01" ...
#>  $ incident_end_date                : POSIXct, format: NA NA ...
#>  $ arrest_date                      : POSIXct, format: "2018-12-31 22:55:00" "2018-12-31 23:28:00" ...
#>  $ law_enforcement_agency           : Factor w/ 499 levels "CHICAGO POLICE DEPT",..: 170 170 170 170 170 170 170 170 170 170 ...
#>  $ unit                             : chr  "District 10 - Ogden" "District 10 - Ogden" "District 25 - Grand Central" "District 25 - Grand Central" ...
#>  $ incident_city                    : chr  "Chicago" "Chicago" "Chicago" "Chicago" ...
#>  $ received_date                    : POSIXct, format: "2019-01-01" "2019-01-01" ...
#>  $ arraignment_date                 : POSIXct, format: "2019-02-05" "2019-02-07" ...
#>  $ updated_offense_category         : chr  "Reckless Discharge of Firearm" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "Reckless Discharge of Firearm" ...
#>  $ charge_count                     : int  1 11 5 2 1 1 1 1 4 1 ...