Initiation Data

The Initiation results data presented here reflects all of the arrests that came through the door of the State's Attorneys Office (SAO). An initiation is how an arrest turns into a “case” in the courts. Most cases are initiated through a process known as felony review, in which SAO attorneys make a decision whether or not to prosecute. Cases may also be indicted by a grand jury or, in narcotics cases, filed directly by law enforcement (labeled "BOND SET (Narcotics)" in this data). Included in this data set are the defendant counts by initiation and year. This data includes felony cases handled by the Criminal, Narcotics, and Special Prosecution Bureaus. It does not include information about cases processed through the Juvenile Justice and Civil Actions Bureaus.

initiation

Format

The data frame intake has 73,029 rows and 27 variables:

case_id

Internal unique identifier for each case

case_participant_id

Internal unique identifier for each person associated with a case

offense_category

Broad offense categories before specific charges are filed on a case

primary_charge

A flag for the top charge, usually the way the case is referred to

charge_id

Internal unique identifier for each charge filed

charge_version_id

Internal unique identifier for each version of a charge associated with charges filed

charge_offense_title

The offense title of the charged offense

chapter

The legal chapter for the charge

act

The legal act for the charge

section

The legal section for the charge

class

The legal class for the charge

aoic

Administrative Office of the Illinois Courts ID for law of the charge

event

The way the charge was brought about

event_date

The date the charges were brought about

age_at_incident

Recorded age at the time of the incident

gender

Recorded gender of the defendant

race

Recorded race of the defendant

incident_begin_date

Date of when the incident began

incident_end_date

Date of when the incident ended (this will be blank for incidents that did not go more than one day)

arrest_date

Date and time of arrest

law_enforcement_agency

Law enforcement agency associated with the arrest

unit

The law enforcement unit associated with the arrest

incident_city

The city where the incident took place

received_date

Date when felony review received the case

arraignment_date

Date of the arraignment

update_offense_category

This field is the offense category for the case updated based upon the top charge for the primary offender. It can differ from the first offense category assigned to the case in part because cases evolve.

charge_count

The charge count of the charged offense

Source

https://datacatalog.cookcountyil.gov/Courts/Initiation/7mck-ehwz

See also

Examples

str(initiation)

#> 'data.frame':	73029 obs. of  27 variables:
#>  $ case_id                 : chr  "129405877581" "129568427364" "129568471166" "129568471166" ...
#>  $ case_participant_id     : chr  "1077995851303" "1079958749626" "1079959687690" "1079959687690" ...
#>  $ offense_category        : chr  "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" ...
#>  $ primary_charge          : chr  "true" "true" "false" "false" ...
#>  $ charge_id               : chr  "2529188058314" "2534063384335" "2534433381475" "2534434268758" ...
#>  $ charge_version_id       : chr  "576544933565" "577688601680" "577775465948" "577775667334" ...
#>  $ charge_offense_title    : chr  "[POSSESSION OF CONTROLLED SUBSTANCE WITH INTENT TO DELIVER/DELIVERY OF A CONTROLLED SUBSTANCE]" "UNLAWFUL USE OR POSSESSION OF A WEAPON BY A FELON" "AGG UUW/LOADED/NO FCCA/FOID" "AGGRAVATED UNLAWFUL USE OF WEAPON" ...
#>  $ chapter                 : chr  "720" "720" "720" "720" ...
#>  $ act                     : chr  "570" "5" "5" "5" ...
#>  $ section                 : chr  "401(c)(2)" "24-1.1(a)" "24-1.6(a)(1)" "24-1.6(a)(1)" ...
#>  $ class                   : chr  "1" "2" "4" "4" ...
#>  $ aoic                    : chr  "5096200" "0012309" "0017785" "0017862" ...
#>  $ event                   : chr  "Indictment" "Preliminary Hearing" "Preliminary Hearing" "Preliminary Hearing" ...
#>  $ event_date              : POSIXct, format: "2019-06-06" "2019-01-15" ...
#>  $ age_at_incident         : int  31 24 20 20 20 20 20 20 20 20 ...
#>  $ gender                  : chr  "Male" "Male" "Male" "Male" ...
#>  $ race                    : chr  "Black" "Black" "Black" "Black" ...
#>  $ incident_begin_date     : POSIXct, format: "2018-11-12" "2019-01-01" ...
#>  $ incident_end_date       : POSIXct, format: NA NA ...
#>  $ arrest_date             : POSIXct, format: "2018-11-12 23:23:00" NA ...
#>  $ law_enforcement_agency  : chr  "CHICAGO PD" "CHICAGO PD" "CHICAGO PD" "CHICAGO PD" ...
#>  $ unit                    : chr  "District 10 - Ogden" "District 3 - Grand Crossing" "District 22 - Morgan Park" "District 22 - Morgan Park" ...
#>  $ incident_city           : chr  "Chicago" "Chicago" "Chicago" "Chicago" ...
#>  $ received_date           : POSIXct, format: "2019-01-05" "2019-01-01" ...
#>  $ arraignment_date        : POSIXct, format: "2019-06-20" NA ...
#>  $ updated_offense_category: chr  "Narcotics" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" "UUW - Unlawful Use of Weapon" ...
#>  $ charge_count            : int  1 1 2 3 4 5 6 7 8 1 ...